refactor(coded-apps): OS-tar starter kit + modularized build pipeline
Address PR #1591 feedback that the build script was too big and that a zip sat
in source control behind a hand-rolled extractor.

- Replace the committed .zip + hand-rolled zip parser with a .tar.gz fixture,
  extracted and created via the OS `tar` (built into Windows 10+, macOS, Linux)
  — no cross-platform archive code to maintain.
- Extract project-lifecycle utilities (versioning, prewarm, sdk-check,
  dev-server teardown) into lifecycle.mjs.
- Split the four use-case flows (build, edit, upgrade, template) out of
  build-dashboard.mjs into flows/*.mjs; the file drops ~2020 → ~1360 lines and
  is now the shared-primitive library + entry-point dispatcher.
- Move all dashboard scripts under assets/scripts/dashboards/ and update the
  doc + test path references.
- Extract the kit by feeding the archive on stdin (`tar -xz -C <dir> -f - <
  <archive>`), not as a path arg: GNU tar (Git Bash on Windows) otherwise reads
  the `C:\` drive colon as a remote host and dies; `--force-local` fixes GNU tar
  but bsdtar (Windows System32, macOS) rejects it, so stdin is the portable form.

Entry-point contract unchanged. 158/158 unit tests pass; oxlint clean; runtime
import cycle resolves.
