+++ skills/uipath-coded-apps/assets/scripts/dashboards/setup-admin-folder.mjs
